+++ apps/web/src/features/organizations/components/metrics/metric.tsx
@@ -55,6 +55,8 @@ export const Metric = (props: MetricProps) => {
             <DeltaBadge
               deltaType={deltaType}
               isIncreasePositive={isIncreasePositive}
+              variant="ghost"
+              px="1"
             >
               {change}%
             </DeltaBadge>