+++ packages/ui-lib/src/data-board/data-board.tsx
@@ -24,6 +24,7 @@ import {
   getFilteredRowModel,
   getGroupedRowModel,
   useReactTable,
+  RowData,
 } from '@tanstack/react-table'
 import {
   HTMLChakraProps,
@@ -95,6 +96,7 @@ export const DataBoard = forwardRef(
     const instance = useReactTable({
       data: React.useMemo(() => data, []),
       columns: React.useMemo(() => columns, []),
+      groupedColumnMode: false,
       onGroupingChange: setGrouping,
       getGroupedRowModel: getGroupedRowModel(),
       getCoreRowModel: getCoreRowModel(),
@@ -117,24 +119,28 @@ export const DataBoard = forwardRef(
     const state = instance.getState()
 
     const rows = instance.getRowModel().rows
+
     const noResults = (state.columnFilters?.length || state.globalFilter) &&
       !rows.length && <NoResultsComponent onReset={onResetFilters} />
 
     const mapItems = React.useCallback(() => {
-      const items: KanbanItems = {}
+      const items: KanbanItems = groupBy
+        ? getColumns(instance.getPreFilteredRowModel().rows, groupBy)
+        : {}
+
       instance.getRowModel().rows.forEach((row) => {
         if (row.getIsGrouped()) {
           items[row.id] = row.subRows.map((subRow) => subRow.id)
         }
       })
       return items
-    }, [groupBy])
+    }, [groupBy, rows])
 
     React.useEffect(() => {
       setItems(mapItems())
-    }, [groupBy])
+    }, [groupBy, rows])
 
-    const [items, setItems] = React.useState(mapItems())
+    const [items, setItems] = React.useState<KanbanItems>({})
 
     const board = ({ columns, items, activeId }: UseKanbanContainerReturn) => {
       return (
@@ -200,13 +206,30 @@ interface BoardCardProps {
   render: (item: Row<any>) => React.ReactNode
 }
 
-const BoardCard = React.memo(function BoardCard({
-  item,
-  render,
-}: BoardCardProps) {
-  return item ? (
-    <KanbanCard key={item.id} id={item.id}>
-      {flexRender(render, item)}
-    </KanbanCard>
-  ) : null
-})
+const BoardCard = React.memo(
+  function BoardCard({ item, render }: BoardCardProps) {
+    return item ? (
+      <KanbanCard key={item.id} id={item.id}>
+        {flexRender(render, item)}
+      </KanbanCard>
+    ) : null
+  },
+  (prevProps, nextProps) => {
+    // Only re-render if the item data has changed.
+    if (prevProps.item?.original !== nextProps.item?.original) {
+      return false
+    }
+    return true
+  },
+)
+
+function getColumns<TData extends RowData>(
+  rows: Row<TData>[],
+  groupBy: string,
+) {
+  return rows.reduce<Record<string, any>>((columns, row) => {
+    const resKey = `${groupBy}:${row.getGroupingValue(groupBy)}`
+    columns[resKey] = []
+    return columns
+  }, {})
+}