+++ saas-ui/kanban-core/src/hooks/use-safe-layout-effect.ts
@@ -1,0 +1,4 @@
+import { isBrowser } from '@chakra-ui/utils'
+import { useEffect, useLayoutEffect } from 'react'
+
+export const useSafeLayoutEffect = isBrowser ? useLayoutEffect : useEffect