+++ saas-ui/pro-theme-glass/src/components/index.ts
@@ -7,17 +7,23 @@ import { menuStyles } from './menu'
 import { toolbarStyles } from './toolbar'
 import { tabsStyles } from './tabs'
 import { tagStyles } from './tag'
+import { tooltipStyles } from './tooltip'
+import { structuredListStyles } from './structured-list'
+import { selectStyles } from './select'
 import Form from './form'
 
 export const components = {
   Badge: badgeStyles,
   Button: buttonStyles,
   Card: cardStyles,
-  NavItem: navItemStyles,
-  NavGroup: navGroupStyles,
   Menu: menuStyles,
-  Toolbar: toolbarStyles,
   Tabs: tabsStyles,
   Tag: tagStyles,
+  Tooltip: tooltipStyles,
+  SuiToolbar: toolbarStyles,
+  SuiNavItem: navItemStyles,
+  SuiNavGroup: navGroupStyles,
+  SuiStructuredList: structuredListStyles,
+  SuiSelect: selectStyles,
   ...Form,
 }