+++ saas-ui/pro-theme-glass/src/components/tabs.ts
@@ -11,7 +11,6 @@ const variantSegments = definePartsStyle((props) => {
     tablist: {
       display: 'flex',
       bg: 'gray.100',
-      h: '100px',
       rounded: 4,
       _dark: {
         bg: 'whiteAlpha.50',
@@ -51,10 +50,15 @@ const variantSegments = definePartsStyle((props) => {
       _last: { rounded: 4 },
       _hover: {
         borderColor: 'transparent',
+        bg: 'transparent',
       },
       _selected: {
-        ...theme.components.Button.variants.secondary(props),
+        ...theme.components.Button.variants.secondary(props)._selected,
+        bg: 'white',
         color: 'currentColor',
+        _dark: {
+          bg: 'gray.700',
+        },
       },
     },
   }